- Fixed bug #135829   Curl library was called with wrong data-type. The bug appeared on   big-endian 64bits only (Such as s390x). - 2.13.1
svn path=/trunk/transfer/; revision=26541

+++ agent-curl/src/CurlAgent.cc
@@ -41,7 +41,7 @@ YCPValue CurlAgent::Get( const char *url, const char *target)
 	YCPMap	   Response;
 	char	*content;
 	double	dsize, totaltime;
-	int		code;
+	long	code;
 
 
 	outfile = fopen(target, "w");
@@ -104,7 +104,7 @@ YCPValue CurlAgent::Get( const char *url, const char *target)
 		}
 		Response->add(YCPString("dsize"), YCPFloat(dsize));
 
-		y2milestone("curl returned -> response code: %d, total size: %g, total time %f, content type: %s\n", code, dsize, totaltime, content);
+		y2milestone("curl returned -> response code: %ld, total size: %g, total time %f, content type: %s\n", code, dsize, totaltime, content);
 		/* always cleanup */
 		curl_easy_cleanup(curl);
